Renamed src/utils to src/lib
