+++ src/lib/auth.js
